:lipstick: Add sweetalerts to add and delete items

+++ src/app/components/cart-app/cart-app.component.ts
@@ -5,6 +5,7 @@ import { CartItem } from '../../models/cartitem';
 import { NavbarComponent } from '../navbar/navbar.component';
 import { Router, RouterOutlet } from '@angular/router';
 import { SharingDataService } from '../../services/sharing-data.service';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'cart-app',
@@ -53,27 +54,54 @@ export class CartAppComponent implements OnInit{
           total: this.total
         }
       });
-    })
+      Swal.fire({
+        title: "Shopping",
+        text: "¡Nuevo producto agregado al carro!",
+        icon: "success"
+      });
+    });
   }
 
   onDeleteCart(): void{
     this.sharingDataService.idProductEventEmitter.subscribe(id => {
-      console.log(id + ' se ha ejecutado el evento: "idProductEventEmitter"')
-      this.items = this.items.filter(item => item.product.id != id);
-      if(this.items.length == 0){
-        sessionStorage.removeItem('cart');
-        sessionStorage.clear();
-      }
-      this.calculateTotal(); 
-      this.saveSession();
-      this.router.navigateByUrl('/', {skipLocationChange: true}).then(()=>{
-        this.router.navigate(['/cart'], {
-          state: {
-            items : this.items,
-            total: this.total
+    
+      Swal.fire({
+        title: "¿Estaá seguro que desea eliminar?",
+        text: "¡Cuidado el item se eliminará del carro de compras!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "¡Si, eliminarlo!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.items = this.items.filter(item => item.product.id != id);
+          if(this.items.length == 0){
+            sessionStorage.removeItem('cart');
+            sessionStorage.clear();
           }
-        });
-      })
+          this.calculateTotal(); 
+          this.saveSession();
+
+          this.router.navigateByUrl('/', {skipLocationChange: true}).then(()=>{
+            this.router.navigate(['/cart'], {
+              state: {
+                items : this.items,
+                total: this.total
+              }
+            });
+          })
+
+          Swal.fire({
+            title: "¡Eliminado!",
+            text: "Se ha eliminado el item del carro de compras",
+            icon: "success"
+          });
+        }
+      });
+
+      
+      
     })
   }
